JBTM-3924 (6.0) Replace synchronized in favor of Reentrant Lock
 in LRAService joinLRA method
