+++ rts/lra/coordinator/src/test/java/io/narayana/lra/coordinator/tools/osb/mbean/ObjStoreBrowserLRATest.java
@@ -5,7 +5,6 @@
 
 package io.narayana.lra.coordinator.tools.osb.mbean;
 
-import com.arjuna.ats.arjuna.common.Uid;
 import com.arjuna.ats.arjuna.common.recoveryPropertyManager;
 import com.arjuna.ats.arjuna.tools.osb.mbean.ActionBean;
 import com.arjuna.ats.arjuna.tools.osb.mbean.LogRecordWrapper;
@@ -90,7 +89,8 @@ public class ObjStoreBrowserLRATest {
 
     @Test
     public void lraMBeanRemoval() throws Exception {
-        LongRunningAction lra = new LongRunningAction(new Uid());
+        String lraUrl = "http://localhost:8080/lra";
+        LongRunningAction lra = LRARecoveryModule.getService().startLRA(lraUrl, null, "client", Long.MAX_VALUE);
         OSEntryBean lraOSEntryBean = null;
         try {
             lra.begin(Long.MAX_VALUE); // Creating the LRA records in the log store.
